chore: organize transactions dir
- rename dir `assets` -> `types`
- move `transactions/assets/xxxx_display` -> `transactions/ux/xxxx_ux`.
- split `assets/types.h` -> `assets/types`(enum) & `assets/assets`.
