+++ src/operations/transactions/types/htlc_lock.h
@@ -16,40 +16,32 @@
 *  limitations under the License.
 ********************************************************************************/
 
-#ifndef ARK_OPERATIONS_TRANSACTION_H
-#define ARK_OPERATIONS_TRANSACTION_H
+#ifndef ARK_OPERATIONS_TRANSACTIONS_ASSETS_HTLC_LOCK_H
+#define ARK_OPERATIONS_TRANSACTIONS_ASSETS_HTLC_LOCK_H
 
+#include <stdbool.h>
 #include <stddef.h>
 #include <stdint.h>
 
 #include "constants.h"
 
-#include "transactions/types/assets.h"
+////////////////////////////////////////////////////////////////////////////////
+
+static const size_t TRANSACTION_TYPE_HTLC_LOCK_SIZE = 66;
 
 ////////////////////////////////////////////////////////////////////////////////
 
-typedef struct transaction_t {
-    uint8_t     header;
-    uint8_t     version;
-    uint8_t     network;
-    uint16_t    type;
-    uint8_t     senderPublicKey[PUBLICKEY_COMPRESSED_LEN];
-    uint64_t    fee;
-    uint8_t     vendorFieldLength;
-    uint8_t     *vendorField;
-    union {
-        struct {  // v2
-            tx_asset_t  asset;
-        };
-        struct {  // Legacy
-            uint8_t     recipientId[ADDRESS_HASH_LEN];
-            uint64_t    amount;
-            size_t      assetOffset;
-            size_t      assetSize;
-            uint8_t     *assetPtr;
-        };
-    };
-} Transaction;
+typedef struct htlc_lock_asset_t {
+    uint64_t    amount;
+    uint8_t     expirationType;
+    uint32_t    expiration;
+    uint8_t     secretHash[HASH_32_LEN];
+    uint8_t     recipientId[ADDRESS_HASH_LEN];
+} HtlcLock;
+
+////////////////////////////////////////////////////////////////////////////////
+
+bool deserializeHtlcLock(HtlcLock *lock, const uint8_t *buffer, size_t size);
 
 ////////////////////////////////////////////////////////////////////////////////
