+++ src/operations/transactions/types/second_signature.h
@@ -16,40 +16,26 @@
 *  limitations under the License.
 ********************************************************************************/
 
-#ifndef ARK_OPERATIONS_TRANSACTION_H
-#define ARK_OPERATIONS_TRANSACTION_H
+#ifndef ARK_OPERATIONS_TRANSACTIONS_TYPES_SECOND_SIGNATURE_H
+#define ARK_OPERATIONS_TRANSACTIONS_TYPES_SECOND_SIGNATURE_H
 
+#include <stdbool.h>
 #include <stddef.h>
 #include <stdint.h>
 
 #include "constants.h"
 
-#include "transactions/types/assets.h"
+////////////////////////////////////////////////////////////////////////////////
+
+typedef struct ss_registration_asset_t {
+    uint8_t publicKey[PUBLICKEY_COMPRESSED_LEN];
+} SecondSignatureRegistration;
 
 ////////////////////////////////////////////////////////////////////////////////
 
-typedef struct transaction_t {
-    uint8_t     header;
-    uint8_t     version;
-    uint8_t     network;
-    uint16_t    type;
-    uint8_t     senderPublicKey[PUBLICKEY_COMPRESSED_LEN];
-    uint64_t    fee;
-    uint8_t     vendorFieldLength;
-    uint8_t     *vendorField;
-    union {
-        struct {  // v2
-            tx_asset_t  asset;
-        };
-        struct {  // Legacy
-            uint8_t     recipientId[ADDRESS_HASH_LEN];
-            uint64_t    amount;
-            size_t      assetOffset;
-            size_t      assetSize;
-            uint8_t     *assetPtr;
-        };
-    };
-} Transaction;
+bool deserializeSecondSignature(SecondSignatureRegistration *registration,
+                                const uint8_t *buffer,
+                                size_t size);
 
 ////////////////////////////////////////////////////////////////////////////////
