+++ src/operations/transactions/types/transfer.h
@@ -16,8 +16,8 @@
 *  limitations under the License.
 ********************************************************************************/
 
-#ifndef ARK_OPERATIONS_TRANSACTION_ASSETS_VOTE_H
-#define ARK_OPERATIONS_TRANSACTION_ASSETS_VOTE_H
+#ifndef ARK_OPERATIONS_TRANSACTIONS_TYPES_TRANSFER_H
+#define ARK_OPERATIONS_TRANSACTIONS_TYPES_TRANSFER_H
 
 #include <stdbool.h>
 #include <stddef.h>
@@ -27,17 +27,19 @@
 
 ////////////////////////////////////////////////////////////////////////////////
 
-static const size_t VOTE_LEN = sizeof(uint8_t) + PUBLICKEY_COMPRESSED_LEN;
+static const size_t TRANSACTION_TYPE_TRANSFER_SIZE = 33;
 
 ////////////////////////////////////////////////////////////////////////////////
 
-typedef struct vote_asset_t {
-    uint8_t data[VOTE_LEN];  // ( '+/-' + publicKey[33])
-} Vote;
+typedef struct transfer_asset_t {
+    uint64_t    amount;
+    uint32_t    expiration;
+    uint8_t     recipientId[ADDRESS_HASH_LEN];
+} Transfer;
 
 ////////////////////////////////////////////////////////////////////////////////
 
-bool deserializeVote(Vote *vote, const uint8_t *buffer, size_t size);
+bool deserializeTransfer(Transfer *transfer, const uint8_t *buffer, size_t size);
 
 ////////////////////////////////////////////////////////////////////////////////
