+++ src/operations/transactions/ux/display_ux.h
@@ -16,26 +16,14 @@
 *  limitations under the License.
 ********************************************************************************/
 
-#ifndef ARK_OPERATIONS_TRANSACTIONS_ASSETS_SECOND_SIGNATURE_H
-#define ARK_OPERATIONS_TRANSACTIONS_ASSETS_SECOND_SIGNATURE_H
+#ifndef ARK_OPERATIONS_TRANSACTIONS_UX_DISPLAY_UX_H
+#define ARK_OPERATIONS_TRANSACTIONS_UX_DISPLAY_UX_H
 
-#include <stdbool.h>
-#include <stddef.h>
-#include <stdint.h>
-
-#include "constants.h"
+#include "transactions/transaction.h"
 
 ////////////////////////////////////////////////////////////////////////////////
 
-typedef struct ss_registration_asset_t {
-    uint8_t publicKey[PUBLICKEY_COMPRESSED_LEN];
-} SecondSignatureRegistration;
-
-////////////////////////////////////////////////////////////////////////////////
-
-bool deserializeSecondSignature(SecondSignatureRegistration *registration,
-                                const uint8_t *buffer,
-                                size_t size);
+void setDisplay(const Transaction *transaction);
 
 ////////////////////////////////////////////////////////////////////////////////
